# api-contract-snapshot.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Promise Integrated System API Contract",
    "version": "1.0.0",
    "description": "Snapshot of the API contract for quality gating and client generation."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/api/admin/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current admin user",
        "description": "Returns the currently authenticated admin user's profile",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current admin user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Admin login",
        "description": "Authenticate an admin user and create a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "Admin username",
                    "example": "admin"
                  },
                  "password": {
                    "type": "string",
                    "description": "Admin password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Account inactive"
          },
          "429": {
            "description": "Too many login attempts"
          }
        }
      }
    },
    "/api/admin/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Admin logout",
        "description": "Destroy the current admin session",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Logged out successfully"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Logout failed"
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}